Refactor time
This commit removes a lot of `System.nanoTime()`s from the `state` code and the `rendering` code,
using in-game time whenever possible, and making it easier to make future fades/transitions.

+++ importer/src/main/kotlin/mardek/importer/particle/ParticleEmitterImporter.kt
@@ -85,7 +85,7 @@ private fun parseEmissionWaves(content: Content, rawEmitter: Map<String, String>
 	var period = FLASH_FRAME * flashPeriod
 	val particlesPerFrame = parseFloat(rawEmitter["ppf"]!!)
 	val particlesPerWave = if (particlesPerFrame < 1f) {
-		period /= particlesPerFrame
+		period /= particlesPerFrame.toDouble()
 		1
 	} else particlesPerFrame.roundToInt()
 	val flashDuration = parseInt(rawEmitter["dur"]!!)
